requirements.txt
# requirements.txt - Python Dependencies for All Services
Django==5.2.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.32.1
requests==2.32.3
httpx==0.28.0
gunicorn==23.0.0
